rename cache-new -> cache -> cache-old
